+++ src/snowflake/cli/_plugins/nativeapp/sf_sql_facade.py
@@ -39,6 +39,7 @@ from snowflake.cli._plugins.nativeapp.sf_facade_exceptions import (
     UserScriptError,
     handle_unclassified_error,
 )
+from snowflake.cli._plugins.stage.manager import StageManager
 from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.constants import ObjectType
 from snowflake.cli.api.errno import (
@@ -243,7 +244,7 @@ class SnowflakeSQLFacade:
     def create_version_in_package(
         self,
         package_name: str,
-        stage_path_to_artifacts: str,
+        path_to_version_directory: str,
         version: str,
         label: str | None = None,
         role: str | None = None,
@@ -251,7 +252,7 @@ class SnowflakeSQLFacade:
         """
         Creates a new version in an existing application package.
         @param package_name: Name of the application package to alter.
-        @param stage_path_to_artifacts: Path to artifacts on the stage to create a version from.
+        @param path_to_version_directory: Path to artifacts on the stage to create a version from.
         @param version: Version name to create.
         @param [Optional] role: Switch to this role while executing create version.
         @param [Optional] label: Label for this version, visible to consumers.
@@ -264,11 +265,12 @@ class SnowflakeSQLFacade:
         with_label_cause = (
             f"\nlabel={to_string_literal(label)}" if label is not None else ""
         )
+        using_clause = StageManager.quote_stage_name(path_to_version_directory)
         add_version_query = dedent(
             f"""\
                 alter application package {package_name}
                     add version {version}
-                    using @{stage_path_to_artifacts}{with_label_cause}
+                    using {using_clause}{with_label_cause}
             """
         )
         with self._use_role_optional(role):
@@ -283,7 +285,7 @@ class SnowflakeSQLFacade:
     def add_patch_to_package_version(
         self,
         package_name: str,
-        stage_path_to_artifacts: str,
+        path_to_version_directory: str,
         version: str,
         patch: int | None = None,
         label: str | None = None,
@@ -292,7 +294,7 @@ class SnowflakeSQLFacade:
         """
         Add a new patch, optionally a custom one, to an existing version in an application package.
         @param package_name: Name of the application package to alter.
-        @param stage_path_to_artifacts: Path to artifacts on the stage to create a version from.
+        @param path_to_version_directory: Path to artifacts on the stage to create a version from.
         @param version: Version name to create.
         @param [Optional] patch: Patch number to create.
         @param [Optional] label: Label for this patch, visible to consumers.
@@ -309,11 +311,12 @@ class SnowflakeSQLFacade:
             f"\nlabel={to_string_literal(label)}" if label is not None else ""
         )
         patch_query = f"{patch}" if patch else ""
+        using_clause = StageManager.quote_stage_name(path_to_version_directory)
         add_patch_query = dedent(
             f"""\
                  alter application package {package_name}
                      add patch {patch_query} for version {version}
-                     using @{stage_path_to_artifacts}{with_label_clause}
+                     using {using_clause}{with_label_clause}
              """
         )
         with self._use_role_optional(role):
@@ -581,7 +584,7 @@ class SnowflakeSQLFacade:
         self,
         name: str,
         install_method: SameAccountInstallMethod,
-        stage_path_to_artifacts: str,
+        path_to_version_directory: str,
         role: str,
         warehouse: str,
         debug_mode: bool | None,
@@ -592,7 +595,7 @@ class SnowflakeSQLFacade:
 
         @param name: Name of the application object
         @param install_method: Method of installing the application
-        @param stage_path_to_artifacts: Path to directory in stage housing the application artifacts
+        @param path_to_version_directory: Path to directory in stage housing the application artifacts
         @param role: Role to use when creating the application and provider-side objects
         @param warehouse: Warehouse which is required to create an application object
         @param debug_mode: Whether to enable debug mode; None means not explicitly enabled or disabled
@@ -607,7 +610,7 @@ class SnowflakeSQLFacade:
 
         with self._use_role_optional(role), self._use_warehouse_optional(warehouse):
             try:
-                using_clause = install_method.using_clause(stage_path_to_artifacts)
+                using_clause = install_method.using_clause(path_to_version_directory)
                 upgrade_cursor = self._sql_executor.execute_query(
                     f"alter application {name} upgrade {using_clause}",
                 )
@@ -677,7 +680,7 @@ class SnowflakeSQLFacade:
         name: str,
         package_name: str,
         install_method: SameAccountInstallMethod,
-        stage_path_to_artifacts: str,
+        path_to_version_directory: str,
         role: str,
         warehouse: str,
         debug_mode: bool | None,
@@ -690,7 +693,7 @@ class SnowflakeSQLFacade:
         @param name: Name of the application object
         @param package_name: Name of the application package to install the application from
         @param install_method: Method of installing the application
-        @param stage_path_to_artifacts: Path to directory in stage housing the application artifacts
+        @param path_to_version_directory: Path to directory in stage housing the application artifacts
         @param role: Role to use when creating the application and provider-side objects
         @param warehouse: Warehouse which is required to create an application object
         @param debug_mode: Whether to enable debug mode; None means not explicitly enabled or disabled
@@ -712,7 +715,7 @@ class SnowflakeSQLFacade:
             )
             authorize_telemetry_clause = f" AUTHORIZE_TELEMETRY_EVENT_SHARING = {str(should_authorize_event_sharing).upper()}"
 
-        using_clause = install_method.using_clause(stage_path_to_artifacts)
+        using_clause = install_method.using_clause(path_to_version_directory)
         with self._use_role_optional(role), self._use_warehouse_optional(warehouse):
             try:
                 create_cursor = self._sql_executor.execute_query(